configura regras de pluralização
