+++ src/rules/rule-m.js
@@ -1,20 +1,21 @@
 'use strict';
 
-const suffixes = [
-  'm',
-];
-
 /*
  * DESCRIÇÃO:
  *   Pluralizar palavras terminadas em /-m/.
  * REGRA:
  *   Trocar /-m/ por /-ns/.
  */
-const handler = function (word) {
-    return word.replace(/m$/, "ns");
+
+const suffixes = [
+    'm',
+];
+
+const handler = (word) => {
+    return word.replace(/m$/, 'ns');
 };
 
 module.exports = {
-  suffixes,
-  handler,
+    suffixes,
+    handler,
 };
